+++ topsecret-split/v1/internal/model/satellite.go
@@ -2,7 +2,7 @@ package model
 
 type Satellite struct {
 	SatelliteName string  `json:"satellite_name"`
-	Distance      float64 `json:"distance"`
+	Distance      float32 `json:"distance"`
 	Message       string  `json:"Message"`
 }
